Use files block instead of distribution
Distributions must follow a file structure. BINARY distributions must be Zip or Tar files https://jreleaser.org/guide/latest/distributions/binary.html
In this case the release contains a PDF file which should be uploaded as a file, not as a distribution.

Moreover, you may skip defining `owner:` and `apiEndpoint:` as these values are taken directly from the git metadata.